updated the humidifier code to properly support i18n
